Add Discord calls-to-action on home page
Added the same Discord iframe previously used here: https://opendi.org/How%20To%20Contribute/#discord to the contribution section on the main page.
Reformatted contribution section to better fit the rest of the page contents, look more structured.

Since this is down at the bottom of the page, added a less-space-consuming call to view contribution guide and join discord (with links) to the initial intro section as well.