GH Pages refactor: json-schema URL
Made changes to prepare to rename the landing-site repo to opendi-org.github.io, so that it forms the base of the organization's GitHub Pages site.
This will allow us to remove the custom subdomains from the roles and user stories site and what is currently the JSON Schema site.
Those subdomain-sites will instead be hosted at opendi.org/<repo-name>/

Updated links to point to opendi.org/api-specification instead of json-schema.opendi.org
Updated wording/formatting on index.md and sidebars.js (and versioned counterparts) to reflect the incorporation of the JSON Schema into the API spec.