GH Pages refactor: Remove old temp API Spec page
Made changes to prepare to rename the landing-site repo to opendi-org.github.io, so that it forms the base of the organization's GitHub Pages site.
This will allow us to remove the custom subdomains from the roles and user stories site and what is currently the JSON Schema site.

Removed the old placeholder page for the API Specification. This was used as a "coming soon" page for the API Specification, which will exist shortly.
Links to this old page now point to `opendi.org/api-specification` instead.
Once the API Specification site draft is accepted and merged into `main`, this link will not need updating. In the interim it will point to the JSON Schema site.