Homepage: Supporters section, USDA acknowledgement
Added a "Supporters" section to the main OpenDI landing page.
For now, this just has an acknowledgement of USDA AFRI support, with the USDA logo above.

As usual, changes in `versioned_docs/version-Live` mirror the changes in `docs/`, and will not apply to the `Live` version of the deployed site until we merge a PR to `main`.