Add "About the Founders" page
Added a new page describing the founders with headshots and bio info that they provided.

Added this page to the main sidebar, and added a link to it in the "Learn More >>" button under the "Who is OpenDI?" section on the landing page.

NOTE: Changes to `versioned_*` mirror other changes to `docs/`, etc. These changes will not apply to the live site until merged to main.